Configer CatigureStore
