+++ src/redux/configureStore.js
@@ -1,0 +1,17 @@
+// configureStore.js 
+
+import {createStore, combineReducers, applyMiddleware} from 'redux';
+import logger from 'redux-logger';
+import booksReducer from './books/books';
+
+const reducer = combineReducers({
+    booksReducer
+    // additional reducers could be added here
+});
+
+const store = createStore(
+    reducer,
+    applyMiddleware(logger)
+);
+
+export default store;
